backend-fastapi-app/plugins/data-table-plugin/backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=1.10.0
python-dotenv>=0.21.0
sqlalchemy>=2.0.0
mysqlclient>=2.1.0